refactor: use index for types
